Assign animations to single properties
Animations are now assigned through single, discrete properties that are generated with a @for loop. This allows me to also control animation-duration on a per-animation basis and lays the ground for a further refactoring in which every config, keyframe and animation class is created in one single loop. Decision pending, though. Lastly, I assigned a negative value to one of the sway factors, so that one of the bubble animations starts on the opposite direction.

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>