Refactor for saner keyframes+animation creation
Now keyframes and animations are created with a loop that iterates a list with all configuration maps. This makes it easier to add and remove configurations in the future and also organizes the outputs css better. When doing so, though, we still need to update the scripts to assign the new classes. That's a good trade-off anyway.

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>