Totally new celebrate() logic!
Celebrate no longer needs a place parameter. We assume that wherever there is a `.bubble-container` we need a celebration, so we look for them and start the thing. Color and quantity optional configuration can be passed by assigning data attributes to the element itself: super sweet. We also have fallback defaults in case no option is specified. Also, the color variable now is passed to the makeBubble() function, so that each new instance maintains the color. This feels somewhat saner and of course it allows to have different colors for different bubble-containers. It still makes sense, instead, to have one stopCelebrating variable shared among them, so that we can stop them all at once which will be the most common usecase.

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>