Add optional `align` setting
Since we are positioning bubbles with `left from 0 to 100`, some of the bubbles will inevitably overflow the container and this could create a unbalanced visual effect when we use a bubble container on the left and another on the right. This is why I introduced a new align option so that this effect could be somehow reduced. Not bulletproof, but better than nothing. For example, what if we want to use a central alignment? Well there we should apply a correction based on the width of the container, but this would add a lot of logic. Not worthy for now.

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>